# user-storage-extension/src/main/resources/openapi/openapi.yaml
---
openapi: 3.1.0
info:
  title: User Storage API
  version: 1.0-SNAPSHOT
components:
  schemas:
    AuthenticateUserBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
    UserEntity:
      type: object
      properties:
        id:
          type: integer
          format: int64
        enabled:
          type: boolean
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        createdAt:
          $ref: "#/components/schemas/Instant"
        password:
          type: string
paths:
  /users:
    get:
      operationId: listUsers
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserEntity"
      summary: List Users
      tags:
      - User Resource
    post:
      operationId: authenticate
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthenticateUserBody"
        required: true
      responses:
        "200":
          description: Authentication successful
        "401":
          description: Authentication failed
      summary: Authenticate
      tags:
      - User Resource
  /users/{username}:
    get:
      operationId: getUserByUsername
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEntity"
      summary: Get User By Username
      tags:
      - User Resource
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value